Make Quadrature constructor definitions terser.
Removing more out-of-line-yet-inline constructors.
