+++ include/quadrature/quadrature_simpson.h
@@ -45,7 +45,17 @@ public:
    */
   explicit
   QSimpson (const unsigned int _dim,
-            const Order o=THIRD);
+            const Order o=THIRD) :
+    QBase(_dim, o)
+  {
+    // explicitly call the init function in 1D since the
+    // other tensor-product rules require this one.
+    // note that EDGE will not be used internally, however
+    // if we called the function with INVALID_ELEM it would try to
+    // be smart and return, thinking it had already done the work.
+    if (_dim == 1)
+      init(EDGE2);
+  }
 
   /**
    * Destructor. Empty.
@@ -67,23 +77,6 @@ private:
   virtual void init_3D (const ElemType _type=INVALID_ELEM,
                         unsigned int p_level=0) libmesh_override;
 };
-
-
-
-// ------------------------------------------------------------
-// QSimpson class members
-inline
-QSimpson::QSimpson(const unsigned int d,
-                   const Order) : QBase(d,THIRD)
-{
-  // explicitly call the init function in 1D since the
-  // other tensor-product rules require this one.
-  // note that EDGE will not be used internally, however
-  // if we called the function with INVALID_ELEM it would try to
-  // be smart and return, thinking it had already done the work.
-  if (_dim == 1)
-    init(EDGE2);
-}
 
 
 } // namespace libMesh
